[Worklist #21420] Ran lint on edit script (details below).
Prior to and after addressing Worklist Item #21420, it was observed that the
tool bar buttons popping in and out during load.  Taking a while to become visible.

Ran JSHint on the edit.js script to see if something stood out.

Testing with the various lint fixes the odd loading behavior wasn't observed locally any longer.
Pushing this up for testing.  This may be what @CainFoool was seeing as at times it took quite
some time for the buttons to load properly.

JSHint Issues Addressed:
* Resolved instance of \"Use the function form of \"use strict\"\"
** Moved use strict declaration to the file function level.
** This fixed various unknown or not defined errors.
* Resolved instances of \"\'const\' is available in ES6 (use 'esversion: 6') or Mozilla JS extensions (use moz).\"
** Switched const to var to avoid using keywords respected by only a subset of browsers.
* Resolved various multiply defined vars.
** Notable callouts being position var and properties var.
** This should fix issues where the same var was being used though the intent of use may have varied.
* Resolved instances of missing semi-colons

PreFix:  51 JSHint issues detected
PostFix: 0 JSHint issues detected.

Changes to be committed:
	modified:   scripts/system/edit.js