[Case 7734][Broken] Moves selection from Tree base to AssetServer::treeView (details below).
There doesn't appear to be a way of triggering the drag behavior via pressAndHold
on the tree at base level as there's no mouse area available to drive the update.

To get around this, this commit moves the selection behavior from the base to the
treeView within AssetServer using its mouse area.

This leads to undesired behavior though:

The issue is that it fights with the other aspects of the row; for example, blocking
clicks of the expand glyphs. Even though the mouse area is set to
propagateComposedEvents, rejecting an event by setting mouse.accepted to false
doesn't seem to work.  Trying to move the mouse area outside of the treeView, in the
event that the propagation failure is due to being parented to the item we want to catch
the rejected event, didn't appear to mitigate the issue.

Changes Committed:
    modified:   interface/resources/qml/controls-uit/Tree.qml
    modified:   interface/resources/qml/hifi/AssetServer.qml