[Case 4315] Shape Color picker dynamically updates entity color (details below).
* The user no longer needs to click the OK button to submit changes to the color;
rather the color of the entity updates as the user manipulates the color wheel and/or
slider.

* Deleted Variant 1 given merger of changes from master which removes that
version of the control from active use here.

Known Issue/TODO(s):

* Color Picker's Color Preview area has a bug where the original/starting color preview
is black as opposed to reflecting the entity's color upon clicking the color picker.
    * Color preview restore color functionality isn't working.  Clicking the initial
color preview doesn't restore the color.

* ColorControl Variant 2 has RGB fields, in addition to the color picker, so the user
needs to close the picker before being able to tweak the individual color channels.  The
picker has a mode which appears to support color wheel, hex, & rgb color tweaks.  Using this
would allow for a more cohesive experience.

Changes Committed:
    modified:   scripts/system/html/css/colpick.css
    modified:   scripts/system/html/js/entityProperties.js