[Case 4315] Fixes some issues with the color picker (details below).
The preferred color control for Shapes utilized as of Case 7049's fix
was ColorControlVariant2 that allowed the user the option of editing the
color channels in addition to using the color picker wheel controls.  This
was done by having the color picker and then external RGB elements meaning
that the player had to close the picker prior to being able to tweak the RGB
channels themselves.

* This commit fixes the rgbhex mode of the color picker so that all color tweaking
  is available to the user without having to close out the picker.
    * Fixes the increment/decrement buttons for the color channel input fields.  So the
      user has the option of those in addition to directly editing the field value via
      keyboard.
    * Fixes new color vs current color preview section in rgbhex mode

NOTE(s):

Currently rgbhex mode is used only by the base color section as a proof of concept.  Based
on feedback the changes may be proliferated to the other color sections.

Changes Committed:
    modified:   scripts/system/html/css/colpick.css
    modified:   scripts/system/html/js/colpick.js
    modified:   scripts/system/html/css/edit-style.css
    modified:   scripts/system/html/entityProperties.html
    modified:   scripts/system/html/js/entityProperties.js