[Case 4315] Fixes color picker issue in HMD mode (details below).
Previously when in HMD mode, using the color picker wasn't viable as
trying select a color in the gradient or hue areas would result in the
page scrolling as opposed to the element specifying a color to use.

This resolves the issue by turning off touch actions for the elements
of the color picker that should have it when using the color picker.

Tested via Vive Headset.

Changes Committed:
    modified:   scripts/system/html/css/colpick.css