[WL21698] A-Frame: Supports a-circle imports (details below).
Amends ShapeEntityItem::setUnscaledDimensions to enforce the flat dimension value
in the event that the height of a circle or quad (plane) is less than the supported
value (0.0001f) in addition to the prior greater than check.  This covers instances
where it makes sense to assign a value of 0 to the height dimension; however, the
shape ultimately rendered and collided as if it had a height of 1 or more.

Changes Committed:
    modified:   libraries/entities/src/AFrameReader.cpp
    modified:   libraries/entities/src/ShapeEntityItem.cpp