[WL21698] A-Frame: a-mixin support, currently limited to a-entity (details below).
This is currently limited to the a-entity types given the need to support them
arose with the recent target goal of importing several items/scenes from altspacevr
which makes heavy use of both a-entity & a-mixin elements. If need be this support
can be expanded at a later time to include the convenience types relied upon when
using more recent versions of the A-Frame framework.

Testing with source of https://altspacevr.github.io/aframe/examples/test/mixin/
    * Minor Alterations to address:
        * Meta element error within the document for parsing
        * Created an entity to utilized the previously unreferenced sphere mixin asset
        * Added variation of color specifications encountered in reference documents for testing.

Changes Committed:
    modified:   libraries/entities/src/AFrameReader.cpp
    modified:   libraries/entities/src/AFrameReader.h