[WL21698] Updating filter post WL21774; however, not much has changed (details below).
Verified that branch has changes from WL21774's PR #12989; however, in order
to allow for most common variants of the extensions both common cases must be
included or filter won't show files with the extension case forms left out though
the only difference is casing.  For example, if a file has as HTML extension as opposed
to html, and the filter only lists html then the file won't be listed in the browser.

Changes Committed:
    modified:   scripts/system/edit.js